+++ .github/scripts/library_import_tests.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
 
-class=$1
-error=$2
+mode=$1
+class=$2
+error=$3
 
-result=$(python3 -c "import sys; sys.path.append('.'); from zabbix_utils import $class; $class()" 2>&1)
+cmd="import sys; sys.path.append('.'); from zabbix_utils import $class; $class()"
+if [ $mode == "async" ]; then
+    cmd="import sys; import asyncio; sys.path.append('.'); from zabbix_utils import $class; exec('async def main():\n    $class()'); asyncio.run(main())"
+fi
+
+result=$(python3 -c "$cmd" 2>&1)
 echo "$result" | grep "$error" >/dev/null || echo "$result" | (python3 "./.github/scripts/telegram_msg.py" && echo "Error")